+++ ACBrFramework/ACBrFramework.Net/ACBrFramework.Net/Sintegra/Registro/Registro 60R/SintegraRegistro60R.cs
@@ -2,11 +2,11 @@ namespace ACBrFramework.Sintegra
 {
 	public sealed class SintegraRegistro60R
 	{
-		public double BaseDeCalculo { get; set; }
+		public decimal BaseDeCalculo { get; set; }
 
-		public double Valor { get; set; }
+		public decimal Valor { get; set; }
 
-		public double Qtd { get; set; }
+		public decimal Qtd { get; set; }
 
 		public string MesAno { get; set; }
